# web/api/server/models/EnvironmentDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EnvironmentDto = {
    /**
     * The name of the environment
     */
    name: string;
    /**
     * The image of the environment
     */
    image: string;
    /**
     * The builder of the environment
     */
    builder: string;
};